refactor: move `useBoard` to its own file
